+++ test/okhi_flutter_test.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-
-  test('getPlatformVersion', () async {
-    // expect(await OkhiFlutter.platformVersion, '42');
-  });
-}